Renamed spacy_dep.py to spacy_wrapper.py and lint fixes
