+++ src/cltk/dependency/spacy_wrapper.py
@@ -1,4 +1,3 @@
-
 import logging
 import subprocess
 import sys
@@ -7,12 +6,8 @@ from typing import Optional
 import spacy
 from spacy.tokens import Doc
 
-from cltk.core.exceptions import (
-    CLTKException,
-    UnknownLanguageError,
-)
+from cltk.core.exceptions import CLTKException, UnknownLanguageError
 from cltk.utils import suppress_stdout
-
 
 LOG = logging.getLogger(__name__)
 LOG.addHandler(logging.NullHandler())
@@ -30,11 +25,13 @@ class SpacyWrapper:
 
     nlps = {}
 
-    def __init__(self,
-                 language: str,
-                 nlp: Optional[spacy.Language] = None,
-                 interactive: bool = True,
-                 silent: bool = False):
+    def __init__(
+        self,
+        language: str,
+        nlp: Optional[spacy.Language] = None,
+        interactive: bool = True,
+        silent: bool = False,
+    ):
         """
 
         :param language: One of the officially supported languages by CLTK. It selects a default model for
@@ -58,7 +55,9 @@ class SpacyWrapper:
                 )
 
             if not self.is_wrapper_available:
-                raise UnknownLanguageError(f"Language {self._language} is not supported by CLTK wrapper of SpaCy.")
+                raise UnknownLanguageError(
+                    f"Language {self._language} is not supported by CLTK wrapper of SpaCy."
+                )
 
             self._select_model()
 
@@ -101,7 +100,8 @@ class SpacyWrapper:
         >>> spacy_wrapper = SpacyWrapper(language="lat")
         >>> latin_nlp = spacy_wrapper.parse(get_example_text("lat"))
 
-        >>> latin_nlp.sents[0]
+        >>> sents = list(latin_nlp.sents)
+        >>> latin_nlp.
 
 
         :param text: Text to analyze.
@@ -109,15 +109,17 @@ class SpacyWrapper:
         """
         if isinstance(self.nlp, spacy.Language):
             return self.nlp(text)
-        raise CLTKException("No spaCy model has been loaded. Is the language supported? "
-                            "Otherwise try to provide a spaCy model with `nlp`argument.")
+        raise CLTKException(
+            "No spaCy model has been loaded. Is the language supported? "
+            "Otherwise try to provide a spaCy model with `nlp`argument."
+        )
 
     def _download_model(self):
-        if self._language in ["la", "lat"]:
-            package = "https://huggingface.co/latincy/la_core_web_md/resolve/main/la_core_web_md-3.5.1/dist/la_core_web_md-3.5.1.tar.gz"
-            subprocess.check_call([sys.executable, "-m", "pip", "install", package])
-        else:
-            raise CLTKException(f"No spaCy model found for language '{self._language}'")
+        # if self._language in ["la", "lat"]:
+        #     package = "https://huggingface.co/latincy/la_core_web_md/resolve/main/la_core_web_md-3.5.1/dist/la_core_web_md-3.5.1.tar.gz"
+        #     subprocess.check_call([sys.executable, "-m", "pip", "install", package])
+        # else:
+        #     raise CLTKException(f"No spaCy model found for language '{self._language}'")
         spacy.cli.download(self.model)
 
     @property
